requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Document processing and OCR
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
langdetect==1.0.9
pdfplumber==0.10.3

# AI/ML and NLP libraries
torch==2.1.2
transformers==4.35.2
sentence-transformers==2.2.2
huggingface-hub==0.19.4
spacy==3.7.2
nltk==3.8.1
numpy==1.26.4
pandas==2.1.4
textblob==0.17.1

# Web scraping and browser automation
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
scrapy==2.11.0
lxml==4.9.3
html5lib==1.1

# Knowledge graph and data visualization
networkx==3.2.1
matplotlib==3.8.2
plotly==5.17.0
rdflib==7.0.0

# Translation and language processing
googletrans==4.0.0rc1
deep-translator==1.11.4
mtranslate==1.8

# Utilities and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.1
python-magic==0.4.27

# Security and validation
cryptography==41.0.7
bcrypt==4.1.3
email-validator==2.1.0

# Optional: Redis for caching (uncomment if needed)
# redis==5.0.1
# aioredis==2.0.1

# Optional: Neo4j for advanced graph database (uncomment if needed)
# neo4j==5.14.1

# Optional: Vector database support (uncomment if needed)
# faiss-cpu==1.7.4
# chromadb==0.4.18

# Optional: Background task processing (uncomment if needed)
# celery==5.3.4
# flower==2.0.1

# Development and testing (optional, uncomment if needed)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2
# black==23.11.0
# flake8==6.1.0
